Use Int8 instead of Int in Agents.jl ForestFire (#63)

+++ ForestFire/Agents/ForestFire.jl
@@ -5,7 +5,7 @@ using Agents, Random
 function forest_fire(rng, density, griddims)
     space = GridSpaceSingle(griddims; periodic = false, metric = :manhattan)
     ## Empty = 0, Green = 1, Burning = 2, Burnt = 3
-    forest = UnremovableABM(Automata, space; rng, properties = (trees = zeros(Int, griddims),))
+    forest = UnremovableABM(Automata, space; rng, properties = (trees = zeros(UInt8, griddims),))
     for I in findall(<(density), rand(abmrng(forest), griddims...))
         forest.trees[I] = I[1] == 1 ? 2 : 1
     end
